chore: refactor export interfaces to separate files for better organization
