+++ src/types/account-id-with-key-pair-object.ts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+
+export interface AccountIdWithKeyPairObject {
+  accountId: string;
+  privateKey: string;
+  publicKey: string;
+}
