+++ src/types/gossip-endpoint.ts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+
+export interface GossipEndpoint {
+  nodeId: number;
+  hostname: string;
+  port: number;
+}
