+++ src/types/node-account-id.ts
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: Apache-2.0
+
+export interface NodeAccountId {
+  accountId: {
+    realm: string;
+    shard: string;
+    accountNum: string;
+  };
+}
