+++ src/types/port-forward-config.ts
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+
+export interface PortForwardConfig {
+  localPort: number;
+  podPort: number;
+}
