+++ src/types/private-key-and-certificate-object.ts
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+
+export interface PrivateKeyAndCertificateObject {
+  privateKeyFile: string;
+  certificateFile: string;
+}
