+++ src/types/release-info.ts
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: Apache-2.0
+
+export interface ReleaseInfo {
+  downloadUrl: string;
+  assetName: string;
+  checksum: string;
+  version: string;
+}
