+++ src/types/service-endpoint.ts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+
+export interface ServiceEndpoint {
+  ipAddressV4?: string;
+  port: number;
+  domainName: string;
+}
