+++ src/types/to-object.ts
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Interface for converting a class to a plain object.
+ */
+export interface ToObject<T> {
+  /**
+   * Converts the class instance to a plain object.
+   *
+   * @returns the plain object representation of the class.
+   */
+  toObject(): T;
+}
